Code refactoring.
Move ActionBlock to cells. Add new attribute for add/delete actions. Create function for initialize actions.
