+++ code/richediting/src/main/java/org/fbme/ide/richediting/adapters/ecc/ECStateController.java
@@ -6,10 +6,8 @@ import jetbrains.mps.nodeEditor.cells.*;
 import jetbrains.mps.openapi.editor.EditorContext;
 import jetbrains.mps.openapi.editor.TextBuilder;
 import org.fbme.ide.iec61499.repository.PlatformElement;
-import org.fbme.ide.richediting.adapters.ecc.cell.ActionCell;
-import org.fbme.ide.richediting.adapters.ecc.cell.AlgorithmCell;
-import org.fbme.ide.richediting.adapters.ecc.cell.OutputCell;
-import org.fbme.ide.richediting.adapters.ecc.cell.StateCell;
+import org.fbme.ide.richediting.adapters.ecc.cell.*;
+import org.fbme.lib.iec61499.declarations.AlgorithmDeclaration;
 import org.fbme.lib.iec61499.ecc.StateAction;
 import org.fbme.lib.iec61499.ecc.StateDeclaration;
 import org.fbme.scenes.controllers.LayoutUtil;
@@ -20,47 +18,30 @@ import org.jetbrains.mps.openapi.model.SNode;
 import java.awt.*;
 import java.util.List;
 import java.util.ArrayList;
-import java.util.Objects;
 import java.util.function.Function;
 
 public class ECStateController implements ComponentController<Point> {
-    private final StateCell myStateNameCell;
-    private final EditorCell_Collection myCellCollection;
-    private final List<AlgorithmBlock> myAlgorithmCells;
-
+    private final EditorContext myContext;
     private final StateDeclaration myState;
-
-    private final boolean myEditable;
+    private final StateCell myStateNameCell;
+    private final SNode myNode;
+    private final EditorCell_Collection myCellCollection;
+    private final List<ActionBlock> myStateActionBlocks;
 
     public ECStateController(EditorContext context, StateDeclaration state) {
+        myContext = context;
         myState = state;
-        myEditable = true;
-        myAlgorithmCells = new ArrayList<>();
-        SNode node = ((PlatformElement) myState).getNode();
-        myCellCollection = createRootCell(context, node);
-        myStateNameCell = createStateCell(context, node);
+        myStateActionBlocks = new ArrayList<>();
+        myNode = ((PlatformElement) state).getNode();
+
+        myCellCollection = createRootCell(myContext, myNode);
+
+        myStateNameCell = StateCell.createStateCell(myContext, myNode, myState);
         myCellCollection.addEditorCell(myStateNameCell);
 
-        for (StateAction action : myState.getActions()) {
-            AlgorithmDeclaration algorithmDeclaration = action.getAlgorithm().getTarget();
-            jetbrains.mps.openapi.editor.cells.EditorCell bodyCell = null;
-            ActionCell actionCell;
-            if (algorithmDeclaration != null) {
-                SNode algorithmNode = ((PlatformElement) algorithmDeclaration.getBody()).getNode();
-                bodyCell = context.getEditorComponent().getUpdater().getCurrentUpdateSession().updateChildNodeCell(algorithmNode);
-                actionCell = new AlgorithmCell(context, node, action, myCellCollection, bodyCell);
-                myCellCollection.addEditorCell(actionCell);
-                myCellCollection.addEditorCell(bodyCell);
-            } else {
-                actionCell = new AlgorithmCell(context, node, action, myCellCollection, null);
-                myCellCollection.addEditorCell(actionCell);
-            }
-
-            ActionCell outputCell = new OutputCell(context, node, action, myCellCollection);
-            myCellCollection.addEditorCell(outputCell);
-            myAlgorithmCells.add(new AlgorithmBlock(actionCell, outputCell, bodyCell));
-        }
+        initializeActions();
         myCellCollection.setBig(true);
+
         relayout();
     }
 
@@ -77,6 +58,28 @@ public class ECStateController implements ComponentController<Point> {
         });
     }
 
+    private void initializeActions() {
+        for (StateAction action : myState.getActions()) {
+            AlgorithmDeclaration algorithmDeclaration = action.getAlgorithm().getTarget();
+            jetbrains.mps.openapi.editor.cells.EditorCell bodyCell = null;
+            ActionCell actionCell;
+            if (algorithmDeclaration != null) {
+                SNode algorithmNode = ((PlatformElement) algorithmDeclaration.getBody()).getNode();
+                bodyCell = myContext.getEditorComponent().getUpdater().getCurrentUpdateSession().updateChildNodeCell(algorithmNode);
+                actionCell = new AlgorithmCell(myContext, myNode, action, myCellCollection, bodyCell, myState);
+                myCellCollection.addEditorCell(actionCell);
+                myCellCollection.addEditorCell(bodyCell);
+            } else {
+                actionCell = new AlgorithmCell(myContext, myNode, action, myCellCollection, null, myState);
+                myCellCollection.addEditorCell(actionCell);
+            }
+
+            ActionCell outputCell = new OutputCell(myContext, myNode, action, myCellCollection, myState);
+            myCellCollection.addEditorCell(outputCell);
+            myStateActionBlocks.add(new ActionBlock(actionCell, outputCell, bodyCell));
+        }
+    }
+
     public void relayout() {
         myStateNameCell.relayout();
 
@@ -84,7 +87,7 @@ public class ECStateController implements ComponentController<Point> {
         int width = myStateNameCell.getWidth();
         int height = getLineSize();
 
-        for (AlgorithmBlock block: myAlgorithmCells) {
+        for (ActionBlock block: myStateActionBlocks) {
             block.relayout();
             width = block.newWidth(width);
             height += block.getHeight(padding);
@@ -94,7 +97,7 @@ public class ECStateController implements ComponentController<Point> {
         myCellCollection.setHeight(height);
 
         myStateNameCell.setWidth(width + ActionCell.ACTIVE_WEIGHT_PADDING);
-        for (AlgorithmBlock block: myAlgorithmCells) {
+        for (ActionBlock block: myStateActionBlocks) {
             block.setWidth(width);
         }
 
@@ -102,9 +105,9 @@ public class ECStateController implements ComponentController<Point> {
         myStateNameCell.moveTo(dx, myCellCollection.getY());
 
         int currentHeight = myStateNameCell.getHeight() + padding;
-        for (AlgorithmBlock algorithmBlock : myAlgorithmCells) {
+        for (ActionBlock ActionBlock : myStateActionBlocks) {
             int dy = myCellCollection.getY();
-            currentHeight = algorithmBlock.moveTo(dx, dy, padding, currentHeight);
+            currentHeight = ActionBlock.moveTo(dx, dy, padding, currentHeight);
         }
     }
 
@@ -114,7 +117,7 @@ public class ECStateController implements ComponentController<Point> {
 
     @Override
     public boolean canStartMoveAt(Point position, int x, int y) {
-        return myEditable;
+        return true;
     }
 
     @NotNull
@@ -156,93 +159,5 @@ public class ECStateController implements ComponentController<Point> {
     @Override
     public void paintTrace(Graphics g, Point position) {
         // do nothing
-    }
-
-    private StateCell createStateCell(EditorContext editorContext, SNode node) {
-        ModelAccessor modelAccessor = new ModelAccessor() {
-            public String getText() {
-                String name = myState.getName();
-                return Objects.equals(name, "") ? null : name;
-            }
-
-            public void setText(String text) {
-                myState.setName(text == null ? "" : text);
-            }
-
-            public boolean isValidText(String text) {
-                return text != null && !text.equals("");
-            }
-        };
-        return new StateCell(editorContext, modelAccessor, node);
-    }
-
-    private static class AlgorithmBlock {
-        private final ActionCell action;
-        private final ActionCell output;
-        private final EditorCell_Collection bodyCell;
-        AlgorithmBlock(ActionCell action, ActionCell output, jetbrains.mps.openapi.editor.cells.EditorCell bodyCell) {
-            this.action = action;
-            this.output = output;
-            if (bodyCell instanceof EditorCell_Collection) {
-                this.bodyCell = (EditorCell_Collection) bodyCell;
-            } else {
-                this.bodyCell = null;
-            }
-        }
-
-        public ActionCell getAction() {
-            return action;
-        }
-
-        public ActionCell getOutput() {
-            return output;
-        }
-
-        public void relayout() {
-            action.relayout();
-            if (bodyCell != null) {
-                bodyCell.relayout();
-                for (jetbrains.mps.openapi.editor.cells.EditorCell editorCell : bodyCell) {
-                    editorCell.relayout();
-                }
-            }
-            output.relayout();
-        }
-
-        public int newWidth(int oldWidth) {
-            int maxTmp = Math.max(action.getWidth(), output.getWidth());
-            if (bodyCell != null) {
-                maxTmp = Math.max(bodyCell.getWidth(), maxTmp);
-            }
-            return Math.max(oldWidth, maxTmp);
-        }
-
-        public void setWidth(int width) {
-            action.setWidth(width);
-            output.setWidth(width);
-            if (bodyCell != null) {
-                bodyCell.setWidth(width);
-            }
-        }
-
-        public int getHeight(int padding) {
-            int x = 0;
-            if (bodyCell != null) {
-                x = bodyCell.getHeight();
-            }
-            return action.getHeight() + output.getHeight() + 3 * padding + x;
-        }
-
-        public int moveTo(int dx, int dy, int padding, int currentHeight) {
-            action.moveTo(dx, dy + currentHeight);
-            currentHeight += (action.getHeight() + padding);
-            if (bodyCell != null) {
-                bodyCell.moveTo(dx, dy + currentHeight);
-                currentHeight += (bodyCell.getHeight() + padding);
-            }
-            output.moveTo(dx, dy + currentHeight);
-            currentHeight += (output.getHeight() + padding);
-            return currentHeight;
-        }
     }
 }
