+++ code/richediting/src/main/java/org/fbme/ide/richediting/adapters/ecc/cell/OutputCell.java
@@ -5,6 +5,7 @@ import jetbrains.mps.openapi.editor.EditorContext;
 import org.fbme.ide.richediting.editor.RichEditorStyleAttributes;
 import org.fbme.lib.iec61499.declarations.EventDeclaration;
 import org.fbme.lib.iec61499.ecc.StateAction;
+import org.fbme.lib.iec61499.ecc.StateDeclaration;
 import org.fbme.lib.iec61499.fbnetwork.PortPath;
 import org.jetbrains.mps.openapi.model.SNode;
 
@@ -13,8 +14,14 @@ import java.awt.*;
 public class OutputCell extends ActionCell {
     private static final Color OUTPUT_COLOR = new Color(235, 221, 185);
 
-    public OutputCell(EditorContext editorContext, SNode node, StateAction action, EditorCell_Collection cellCollection) {
-        super(editorContext, node,OUTPUT_COLOR, action, cellCollection);
+    public OutputCell(
+            EditorContext editorContext,
+            SNode node,
+            StateAction action,
+            EditorCell_Collection cellCollection,
+            StateDeclaration state
+    ) {
+        super(editorContext, node,OUTPUT_COLOR, action, cellCollection, state);
         getStyle().set(RichEditorStyleAttributes.OUTPUTS, action);
     }
